--- a/src/NexusMods.DataModel.RateLimiting/Extensions/ResourceExtensions.cs
+++ b/src/NexusMods.DataModel.RateLimiting/Extensions/ResourceExtensions.cs
@@ -19,7 +19,7 @@ public static class ResourceExtensions
     /// <typeparam name="TItem">Type of item returned from user function.</typeparam>
     /// <returns></returns>
     public static IAsyncEnumerable<TItem> ForEachFileAsync<TResource, TItem>(this IResource<TResource, Size> resource,
-        AbsolutePath root, Func<IJob<Size>, FileEntry, ValueTask<TItem>> fn,
+        AbsolutePath root, Func<IJob<Size>, IFileEntry, ValueTask<TItem>> fn,
         CancellationToken? token = null, string jobName = "Processing Files")
     {
         return ForEachFileAsync(resource, new[] { root }, fn, token, jobName);
@@ -41,7 +41,7 @@ public static class ResourceExtensions
     ///    will only get results once a slice [group] of files are finished processing.
     /// </remarks>
     public static async IAsyncEnumerable<TItem> ForEachFileAsync<TResource, TItem>(this IResource<TResource, Size> resource,
-        IEnumerable<AbsolutePath> roots, Func<IJob<Size>, FileEntry, ValueTask<TItem>> fn,
+        IEnumerable<AbsolutePath> roots, Func<IJob<Size>, IFileEntry, ValueTask<TItem>> fn,
         CancellationToken? token = null, string jobName = "Processing Files")
     {
         token ??= CancellationToken.None;