--- a/src/NexusMods.Paths/AbsolutePath.cs
+++ b/src/NexusMods.Paths/AbsolutePath.cs
@@ -74,11 +74,11 @@ public readonly partial struct AbsolutePath : IEquatable<AbsolutePath>, IPath
     /// file name.
     /// </summary>
     /// <param name="directoryPath"></param>
-    /// <param name="fullPath"></param>
+    /// <param name="fileName"></param>
     /// <param name="fileSystem"></param>
     /// <returns></returns>
-    public static AbsolutePath FromDirectoryAndFileName(string directoryPath, string fullPath, IFileSystem? fileSystem = null)
-        => new(directoryPath, fullPath, fileSystem);
+    public static AbsolutePath FromDirectoryAndFileName(string directoryPath, string fileName, IFileSystem? fileSystem = null)
+        => new(directoryPath, fileName, fileSystem);
 
     /// <summary>
     /// Create a new <see cref="AbsolutePath"/> from an existing full path.