… benchmarks/NexusMods.Benchmarks/Benchmarks/EnumerateFiles.cs
… src/NexusMods.Paths/FileEntry.cs
… src/NexusMods.Paths/FileSystemAbstraction/BaseFileSystem.cs
… src/NexusMods.Paths/FileSystemAbstraction/IFileSystem.cs
… src/NexusMods.Paths/FileSystemAbstraction/InMemoryFileSystem/InMemoryFileSystem.cs
… src/NexusMods.Paths/FileSystemAbstraction/RealFileSystem/DirectoriesEnumerator.cs
… src/NexusMods.Paths/FileSystemAbstraction/RealFileSystem/FileSystem.cs
… src/NexusMods.Paths/FileSystemAbstraction/RealFileSystem/FilesEnumerator.cs
… src/NexusMods.Paths/FileSystemAbstraction/RealFileSystem/FilesEnumeratorEx.cs
… src/NexusMods.Paths/NexusMods.Paths.csproj
… src/NexusMods.Paths/NexusMods.Paths.csproj.DotSettings
… src/NexusMods.Paths/Services.cs
… src/NexusMods.Paths/TemporaryFileManager.cs
… src/NexusMods.Paths/Utilities/EnumeratorHelpers.cs
… src/NexusMods.Paths/Utilities/Internal/Enumerators/Common.cs
… src/NexusMods.Paths/Utilities/Internal/Enumerators/DirectoriesEnumerator.cs
… src/NexusMods.Paths/Utilities/Internal/Enumerators/FilesEnumerator.cs
… src/NexusMods.Paths/Utilities/Internal/Enumerators/FilesEnumeratorEx.cs
… tests/NexusMods.DataModel.Tests/FileHashCacheTests.cs
… tests/NexusMods.Paths.Tests/ExtensionTests.cs
… tests/NexusMods.Paths.Tests/FileSystem/FileSystemTests.cs
… tests/NexusMods.Paths.Tests/FileSystem/InMemoryFileSystemTests.cs
… tests/NexusMods.Paths.Tests/New/AbsolutePathTests/ConstructorTests.cs
… tests/NexusMods.Paths.Tests/New/AbsolutePathTests/FromFullPathTests.cs
… tests/NexusMods.Paths.Tests/New/AbsolutePathTests/GetFullPathTest.cs