Resolves #189

Changes:
- `AbsolutePath` is now marked as `readonly`
- `AbsolutePath` requires a rooted path and doesn't accept relative paths
- `AbsolutePath.Directory` is guaranteed to not be null
- `AbsolutePath.GetFullPath` and all overloads use the same function internally
- removed `AbsolutePath.GetFullPathWithSeparator`
- added `AddFileSystem` which adds a singleton implementation for `IFileSystem`
- marked methods which redirect to `_fileSystem` as obsolete, new code should use `IFileSystem` directly
- removed `FileEntry` and replaced all usages with `IFileEntry`
- added and implemented file and directory enumerator methods to `IFileSystem`
- more tests